pyproject.toml
[project]
name = "stsmoddergui"
version = "0.1.0"
description = "Slay the Spire modding utilities with JPype BaseMod wrapper"
authors = [{name = "stsmoddergui Team"}]
requires-python = ">=3.11"
dependencies = [
  "jpype1>=1.4.1",
]

[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"
